restructure files in tests
