+++ src/typed-method-types/workflows/triggers/tests/event_test.ts
@@ -1,31 +1,39 @@
 import { assertEquals } from "https://deno.land/std@0.99.0/testing/asserts.ts";
-import { ShortcutTrigger } from "./shortcut.ts";
-import { TriggerTypes } from "./mod.ts";
+import { EventTrigger } from "../event.ts";
+import { TriggerTypes } from "../mod.ts";
+import { SlackAPI } from "../../../../mod.ts";
 import * as mf from "https://deno.land/x/mock_fetch@0.3.0/mod.ts";
-import { SlackAPI } from "../../../mod.ts";
-import { shortcut_response } from "./fixtures/sample_responses.ts";
+import { event_response } from "./fixtures/sample_responses.ts";
 
-Deno.test("Shortcut Triggers can set the type using the string", () => {
-  const shortcut: ShortcutTrigger = {
-    type: "shortcut",
+Deno.test("Event Triggers can set the type using the string", () => {
+  const event: EventTrigger = {
+    type: "event",
     name: "test",
     workflow: "#/workflows/example",
     inputs: {},
+    event: {
+      event_type: "reaction_added",
+      channel_ids: ["C013ZG3K41Z"],
+    },
   };
-  assertEquals(shortcut.type, TriggerTypes.Shortcut);
+  assertEquals(event.type, TriggerTypes.Event);
 });
 
-Deno.test("Shortcut Triggers can set the type using the TriggerTypes object", () => {
-  const shortcut: ShortcutTrigger = {
-    type: TriggerTypes.Shortcut,
+Deno.test("Event Triggers can set the type using the TriggerTypes object", () => {
+  const event: EventTrigger = {
+    type: TriggerTypes.Event,
     name: "test",
     workflow: "#/workflows/example",
     inputs: {},
+    event: {
+      event_type: "reaction_added",
+      channel_ids: ["C013ZG3K41Z"],
+    },
   };
-  assertEquals(shortcut.type, TriggerTypes.Shortcut);
+  assertEquals(event.type, TriggerTypes.Event);
 });
 
-Deno.test("Mock call for shortcut", async (t) => {
+Deno.test("Mock call for event", async (t) => {
   mf.install(); // mock out calls to `fetch`
 
   await t.step("instantiated with default API URL", async (t) => {
@@ -36,7 +44,7 @@ Deno.test("Mock call for shortcut", async (t) => {
       assertEquals(typeof client.workflows.triggers.update, "function");
     });
 
-    await t.step("shortcut responses", async (t) => {
+    await t.step("apiCall method", async (t) => {
       await t.step(
         "should return successful response JSON on create",
         async () => {
@@ -45,27 +53,32 @@ Deno.test("Mock call for shortcut", async (t) => {
               req.url,
               "https://slack.com/api/workflows.triggers.create",
             );
-            return new Response(JSON.stringify(shortcut_response));
+            return new Response(JSON.stringify(event_response));
           });
 
           const res = await await client.workflows.triggers.create({
             name: "TEST",
-            type: "shortcut",
+            type: "event",
             workflow: "#/workflows/reverse_workflow",
             inputs: {
               a_string: {
                 value: "string",
               },
             },
+            event: {
+              event_type: "reaction_added",
+              channel_ids: ["C013ZG3K41Z"],
+            },
           });
           assertEquals(res.ok, true);
           if (res.ok) {
-            assertEquals(res.trigger, shortcut_response.trigger);
+            assertEquals(res.trigger, event_response.trigger);
             assertEquals<string>(
-              res.trigger?.shortcut_url,
-              shortcut_response.trigger.shortcut_url,
+              res.trigger?.event_type,
+              event_response.trigger.event_type,
             );
           }
+
           mf.reset();
         },
       );
@@ -79,12 +92,12 @@ Deno.test("Mock call for shortcut", async (t) => {
             req.url,
             "https://slack.com/api/workflows.triggers.update",
           );
-          return new Response(JSON.stringify(shortcut_response));
+          return new Response(JSON.stringify(event_response));
         });
 
         const res = await await client.workflows.triggers.update({
           name: "TEST",
-          type: "shortcut",
+          type: "event",
           trigger_id: "123",
           workflow: "#/workflows/reverse_workflow",
           inputs: {
@@ -92,13 +105,17 @@ Deno.test("Mock call for shortcut", async (t) => {
               value: "string",
             },
           },
+          event: {
+            event_type: "reaction_added",
+            channel_ids: ["C013ZG3K41Z"],
+          },
         });
         assertEquals(res.ok, true);
         if (res.ok) {
-          assertEquals(res.trigger, shortcut_response.trigger);
+          assertEquals(res.trigger, event_response.trigger);
           assertEquals<string>(
-            res.trigger?.shortcut_url,
-            shortcut_response.trigger.shortcut_url,
+            res.trigger?.event_type,
+            event_response.trigger.event_type,
           );
         }
